elroy/web_api/requirements.txt
# Web API dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
